Create and integrate mock brokerage
To test the features relating to brokerage integration, the actual brokerage servers were used for validation. Although harder to control, this was fine for requests that requested basic information. This method will not work sufficiently for currency transactions due to fees being applied on every transaction. To fix this, a mock brokerage system was created to simulate the brokerage responses. This will allow tests to be conducted as many times as necessary without the need to pay transaction fees for each test.